[explorer/rest] refactor: move NEM models into chain-specific package
