+++ explorer/rest/rest/db/NemDatabase.py
@@ -5,17 +5,17 @@ from symbolchain.CryptoTypes import PublicKey
 from symbolchain.nc import TransactionType
 from symbolchain.nem.Network import Address
 
-from rest.model.Account import AccountView
-from rest.model.Block import BlockView
-from rest.model.Mosaic import MosaicRichListView, MosaicView
-from rest.model.Namespace import NamespaceView
-from rest.model.Statistic import (
+from rest.model.nem.Account import AccountView
+from rest.model.nem.Block import BlockView
+from rest.model.nem.Mosaic import MosaicRichListView, MosaicView
+from rest.model.nem.Namespace import NamespaceView
+from rest.model.nem.Statistic import (
 	StatisticAccountView,
 	StatisticTransactionDateRangeDataView,
 	StatisticTransactionDateRangeView,
 	StatisticTransactionView
 )
-from rest.model.Transaction import TransactionRecord, TransactionView
+from rest.model.nem.Transaction import TransactionRecord, TransactionView
 
 from .DatabaseConnection import DatabaseConnectionPool
