+++ explorer/rest/rest/model/nem/Namespace.py
